requirements.txt
<<<<<<< HEAD
beautifulsoup4==4.9.1
requests==2.24.0
fastai==2.0.15
pathlib==1.0.1
selenium==3.141.0
pillow==7.2.0
numpy==1.18.5
wget==3.2
torchvision==0.7.0
pandas==1.1.1
matplotlib==3.3.1
opencv-python==4.4.0.42
scikit-learn==0.23.2
tensorflow==2.3.0
tensorboard==2.3.0
tqdm==4.49.0
streamlit==0.66.0
seaborn==0.11.0
Keras==2.3.1
=======
pandas==1.1.1
PyDrive==1.3.1
scikit-learn==0.23.2
seaborn==0.11.0
tensorflow==2.3.0
tqdm==4.48.2
>>>>>>> 913a94141ac68c93faf0fa104e453dedc6db04e6
